requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.10
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
beautifulsoup4>=4.13.4
requests>=2.32.4
pdfminer.six>=20250506
openai>=1.90.0
sendgrid>=6.12.4
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
schedule>=1.2.0
